+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Documented use of numpy docstrings.
 - Helper utilities to take callables with numpy style docstrings and
   create config classes out of them using `make_config`.
+- File listing endpoint to HTTP service.
 ### Changed
 - All instances of `src_url` changed to `key`.
 - Moved from tensorflow 1 to tensorflow 2.